bugfix: prior to this, sometimes tokenize(detokenize(tokens)) != tokens
Fix: make tokenization independent of sop names.

As a consequence of this fix, different compiled weights
can correspond to the same tokenized program,
because the weights depend on the order of sop names (via allocation of
the residual stream to sops). To avoid this, we ensure that detokinize
renames all variables to the names used in tokenization.

Thus
1. if tokens = tokenize(program), compile(program) is not always
equal compile(detokenize(tokens)).
2. when you want a unique set of weights per tokenized program, use
compile(detokenize(tokens)), and NOT compile(program), since the
detokenized model is always the same.